+++ src/Structures/Relationship.php
@@ -3,7 +3,7 @@
 namespace Exonet\Api\Structures;
 
 /**
- * The Relationship Class represents a relationship of an ApiResource with a predefined request to get the related
+ * The Relationship Class represents a relationship of an Resource with a predefined request to get the related
  * resource identifiers.
  */
 class Relationship extends Relation