+++ src/Structures/Resource.php
@@ -7,18 +7,20 @@ namespace Exonet\Api\Structures;
 use Exonet\Api\Exceptions\ExonetApiException;
 
 /**
- * An ApiResource represents a single resource that is retrieved from the API and allows easy access to its attributes
+ * An Resource represents a single resource that is retrieved from the API and allows easy access to its attributes
  * and relations.
  */
-class ApiResource extends ApiResourceIdentifier
+class Resource extends ResourceIdentifier
 {
     /**
      * @var mixed[] The attributes for this resource.
      */
     private $attributes = [];
 
+    private $changedAttributes = [];
+
     /**
-     * ApiResource constructor.
+     * Resource constructor.
      *
      * @param string         $type     The resource type.
      * @param mixed[]|string $contents The contents of the resource, as (already decoded) array or encoded JSON.
@@ -26,6 +28,12 @@ class ApiResource extends ApiResourceIdentifier
     public function __construct(string $type, $contents = [])
     {
         $data = is_array($contents) ? $contents : json_decode($contents, true)['data'];
+
+        // If decoding of JSON has failed, assume the $contents is a hashid.
+        if ($data === null && json_last_error() !== JSON_ERROR_NONE) {
+            $data['id'] = $contents;
+        }
+
         parent::__construct(
             $type,
             $data['id'] ?? null
@@ -46,12 +54,16 @@ class ApiResource extends ApiResourceIdentifier
      * @param string $attributeName The name of the attribute to get.
      * @param mixed  $newValue      The new attribute value.
      *
-     * @return mixed The value of the attribute.
+     * @return mixed The value of the attribute or the ApiResource class when setting an attribute.
      */
     public function attribute($attributeName, $newValue = null)
     {
-        if ($newValue) {
+        // If there are two arguments given, set the value.
+        if (func_num_args() === 2) {
             $this->attributes[$attributeName] = $newValue;
+            $this->changedAttributes[] = $attributeName;
+
+            return $this;
         }
 
         if (!array_key_exists($attributeName, $this->attributes)) {
@@ -72,6 +84,34 @@ class ApiResource extends ApiResourceIdentifier
     }
 
     /**
+     * Patch this resource to the API.
+     *
+     * @return bool True when the patch has succeeded.
+     */
+    public function patch() : bool
+    {
+        if (!empty($this->changedAttributes)) {
+            $this->request->patch($this->id(), $this->toJson(true));
+        }
+
+        if (!empty($this->changedRelationships)) {
+            $relations = $this->toJson(false, true);
+            foreach($relations['data']['relationships'] as $relationName => $relationData) {
+                $this->request->patch($this->id().'/relationships/'.$relationName, $relationData);
+            }
+        }
+
+        return true;
+    }
+
+    public function resetChangedAttributes()
+    {
+        $this->changedAttributes = [];
+
+        return $this;
+    }
+
+    /**
      * Parse relationships found in the resource into related resource identifiers.
      *
      * @param string[] $relationships The relationship data.
@@ -87,11 +127,11 @@ class ApiResource extends ApiResourceIdentifier
 
             if (isset($relation['data']['type'])) {
                 $relationship->setResourceIdentifiers(
-                    new ApiResourceIdentifier($relation['data']['type'], $relation['data']['id'])
+                    new ResourceIdentifier($relation['data']['type'], $relation['data']['id'])
                 );
             } elseif (!empty($relation['data'])) {
                 $relationship->setResourceIdentifiers(
-                    new ApiResourceSet($relation)
+                    new ResourceSet($relation)
                 );
             }
 
@@ -106,7 +146,7 @@ class ApiResource extends ApiResourceIdentifier
      *
      * @return array|null Array that can be used as json.
      */
-    private function toJson() : array
+    private function toJson(bool $onlyChangedAttributes = false, $onlyChangedRelations = false) : array
     {
         $json = [
             'data' => [
@@ -120,15 +160,27 @@ class ApiResource extends ApiResourceIdentifier
         }
 
         // Set the attributes in the json.
-        array_walk($this->attributes, function ($attributeValue, $attributeName) use (&$json) {
-            $json['data']['attributes'][$attributeName] = $attributeValue;
-        });
+        if ($this->attributes && $onlyChangedRelations === false) {
+            array_walk(
+                $this->attributes,
+                function ($attributeValue, $attributeName) use ($onlyChangedAttributes, &$json) {
+                    if ($onlyChangedAttributes === false || in_array($attributeName, $this->changedAttributes, true)) {
+                        $json['data']['attributes'][$attributeName] = $attributeValue;
+                    }
+                }
+            );
+        }
 
         // Set relations.
-        if ($this->relationships) {
-            array_walk($this->relationships, function (Relationship $relation, $name) use (&$json) {
-                $json['data']['relationships'][$name]['data'] = $relation->toJson();
-            });
+        if ($this->relationships && $onlyChangedAttributes === false) {
+            array_walk(
+                $this->relationships,
+                function (Relationship $relation, $name) use ($onlyChangedRelations, &$json) {
+                    if ($onlyChangedRelations === false || in_array($name, $this->changedRelationships, true)) {
+                        $json['data']['relationships'][$name]['data'] = $relation->toJson();
+                    }
+                }
+            );
         }
 
         return $json;