+++ CHANGELOG.md
@@ -20,7 +20,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 ## [v1.0.0](https://github.com/exonet/exonet-api-php/releases/tag/v1.0.0) - 2019-04-29
 [Compare v0.2.0 - v1.0.0](https://github.com/exonet/backend/compare/v0.2.0...v1.0.0)
 ### Breaking
-- The public property `type` in the `ApiResource` class has been renamed to `resourceType` in order not to conflict with the DNS record resource, which has a `type` attribute.
+- The public property `type` in the `Resource` class has been renamed to `resourceType` in order not to conflict with the DNS record resource, which has a `type` attribute.
 
 ### Added
 - Two examples for DNS zones and records.
@@ -29,7 +29,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 [Compare v0.1.0 - v0.2.0](https://github.com/exonet/backend/compare/v0.1.0...v0.2.0)
 ### Added
 - Ready to use examples to get ticket details.
-- The ApiResourceSet now supports ArrayAccess.
+- The ResourceSet now supports ArrayAccess.
 
 ## [v0.1.0](https://github.com/exonet/exonet-api-php/releases/tag/v0.1.0) - 2018-03-08
 ### Added